requirements.txt
asgiref==3.2.3
astroid==2.2.5
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
django-auditlog==0.4.7
django-crispy-forms==1.9.1
django-filter==2.3.0
django-jsonfield==1.4.0
django-tailwind==0.9.0
docutils==0.16
futures==3.1.1
idna==2.10
isort==4.3.21
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==7.1.1
Pygments==2.6.1
pylint==2.3.1
python-dateutil==2.6.0
pytz==2019.3
requests==2.24.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.10
wrapt==1.11.2
